split Game page in two components
